Changed return type
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1531 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/include/filemanager.h
@@ -130,8 +130,8 @@ public:
 	static void getFilename(string const &path, string& filename);
 
   /*! Inherithed from Stream. */
-  virtual int read(char* buffer, int len);
-  virtual int write(char* buffer, int len);
+  virtual u_long read(char* buffer, int len);
+  virtual u_long write(char* buffer, int len);
 
 };
 #endif
